# core/src/main/resources/openapi/openapi-spec.yml
openapi: 3.0.3
info:
  title: Task by Artur Jarosz
  description: "TASK app openAPI specification"
  version: 0.6.0-SNAPSHOT
...
paths:
...
  /supervisions:
    post:
      tags:
        - supervision
      summary: Create supervision
      operationId: createSupervision
      requestBody:
        description: New supervision data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Supervision"
      responses:
        201:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supervision"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
  /supervisions/{supervisionId}:
    put:
      tags:
        - supervision
      summary: Update supervision
      operationId: updateSupervision
      parameters:
        - name: supervisionId
          in: path
          description: "A unique identifier of the supervision"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated supervision data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Supervision"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supervision"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
    delete:
      tags:
        - supervision
      summary: Delete supervision
      operationId: deleteSupervision
      parameters:
        - name: supervisionId
          in: path
          description: "A unique identifier of the supervision"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Supervision successfully deleted
          content: { }
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
    get:
      tags:
        - supervision
      summary: Get supervision
      operationId: getSupervision
      parameters:
        - name: supervisionId
          in: path
          description: "A unique identifier of the supervision"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supervision"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
  /supervisions/{supervisionId}/supervision-visits:
    post:
      tags:
        - supervision
      summary: Create supervision visit
      operationId: createSupervisionVisit
      parameters:
        - name: supervisionId
          in: path
          description: "A unique identifier of the supervision"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: New supervision visit data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupervisionVisit"
      responses:
        201:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupervisionVisit"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
  /supervisions/{supervisionId}/supervision-visits/{supervisionVisitId}:
    put:
      tags:
        - supervision
      summary: Update supervision visit
      operationId: updateSupervisionVisit
      parameters:
        - name: supervisionId
          in: path
          description: "A unique identifier of the supervision"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: supervisionVisitId
          in: path
          description: "A unique identifier of the supervision visit"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: Updated supervision visit data
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupervisionVisit"
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupervisionVisit"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
    get:
      tags:
        - supervision
      summary: Get supervision visit
      operationId: getSupervisionVisit
      parameters:
        - name: supervisionId
          in: path
          description: "A unique identifier of the supervision"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: supervisionVisitId
          in: path
          description: "A unique identifier of the supervision visit"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupervisionVisit"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
    delete:
      tags:
        - supervision
      summary: Delete supervision visit
      operationId: deleteSupervisionVisit
      parameters:
        - name: supervisionId
          in: path
          description: "A unique identifier of the supervision"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
        - name: supervisionVisitId
          in: path
          description: "A unique identifier of the supervision visit"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Successful operation
          content:
            { }
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
...
  /system-parameters:
    get:
      tags:
        - systemParameter
      summary: Get system parameters
      operationId: getSystemParameters
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemParameters"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
  /system-parameters/{systemParameterId}:
    put:
      tags:
        - systemParameter
      summary: Update a system parameter with given supplierId
      operationId: updateSystemParameter
      parameters:
        - name: systemParameterId
          in: path
          description: "A unique identifier of the system parameter"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      requestBody:
        description: Data of updated system parameter
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemParameter"
      responses:
        200:
          description: System parameter successfully updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemParameter"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
    get:
      tags:
        - systemParameter
      summary: Get a system parameter
      operationId: getSystemParameter
      parameters:
        - name: systemParameterId
          in: path
          description: "A unique identifier of the system parameter"
          required: true
          style: simple
          explode: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemParameter"
        400:
          description: Validator error message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidatorError"
components:
  schemas:
...
    Supervision:
      type: object
      properties:
        id:
          type: integer
          format: int64
        projectId:
          type: integer
          format: int64
        hasInvoice:
          type: boolean
        baseNetRate:
          type: number
        hourlyNetRate:
          type: number
        visitNetRate:
          type: number
        hoursCount:
          type: integer
        value:
          type: number
        note:
          type: string
    SupervisionVisit:
      type: object
      properties:
        id:
          type: integer
          format: int64
        supervisionId:
          type: integer
          format: int64
        dateOfVisit:
          type: string
          format: date
        payable:
          type: boolean
        hoursCount:
          type: integer
...
    SystemParameter:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
        value:
          type: string
        defaultValue:
          type: string
        singleValue:
          type: boolean
    SystemParameters:
      type: array
      items:
        $ref: "#/components/schemas/SystemParameter"
    ValidatorError:
      type: object
      properties:
        message:
          type: string
...